fix whitespace handling: treat ONLY the following characters as whitespace: NUL, TAB, LF, CR, SPC Alternatively, there is commented out macro code in quakedef.h to accept any chars in 0..32 as whitespace. Previously: 0..32 and 128..255 (due to signed char)
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@8606 d7cf8633-e32d-0410-b094-e92efae38249

+++ sys_linux.c
@@ -68,7 +68,7 @@ void Sys_PrintToTerminal(const char *text)
 #endif
 	while(*text)
 	{
-		int written = (int)write(1, text, (int)strlen(text));
+		ssize_t written = write(1, text, strlen(text));
 		if(written <= 0)
 			break; // sorry, I cannot do anything about this error - without an output
 		text += written;
@@ -87,7 +87,7 @@ double Sys_DoubleTime (void)
 	if(sys_usenoclockbutbenchmark.integer)
 	{
 		benchmark_time += 1;
-		return benchmark_time / 1e6;
+		return ((double) benchmark_time) / 1e6;
 	}
 #ifdef WIN32
 #include <mmsystem.h>
